feat(BREAKING CHANGE): implement if- and while- statement parsing
